Add preLaunchTask handling to
