+++ src/debugger/utils.ts
@@ -46,3 +46,20 @@ export async function getPtvsdInjectCommand(host: string, port: number, pid: num
     }
     throw new Error("Failed to retrieve ptvsd from Python extension!");
 }
+
+export async function launchFirstTaskMatchingName(name: string) {
+    if (typeof(name) !== "string") {
+        return;
+    }
+    return vscode.tasks.fetchTasks()
+        .then(tasks => {
+            const taskToExecute = tasks.filter((task, i) => {
+                return task.name === name;
+            });
+            
+            if (taskToExecute.length === 0) {
+                throw new Error(`Pre-launch task ${name} not found`);
+            }
+            return vscode.tasks.executeTask(taskToExecute[0]);
+        });
+}
